Add shell and login kwargs to shell()
Allow for the shell to be specified by passing the 'shell' kwarg to the
shell() command with a string specifying the shell to use. Defaults to
"bash".

Allow for a login shell to be specified by passing the 'login' kwarg to
the shell() command with either True or False. Defaults to "False".

Signed-off-by: Nick Travers <n.e.travers@gmail.com>